requirements_txt.txt
# Schwab AI Trading System - Python Dependencies

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# HTTP Client & API
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# WebSocket Support
websockets==12.0

# Authentication & Security
cryptography==41.0.8
keyring==24.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# AI/ML & Deep Learning
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
scikit-learn==1.3.2
numpy==1.24.4
scipy==1.11.4
optuna==3.4.0

# Data Processing & Analysis
pandas==2.1.3
polars==0.19.19
pyarrow==14.0.1
openpyxml==3.1.2

# Technical Analysis
TA-Lib==0.4.28
ta==0.10.2
yfinance==0.2.22
alpha-vantage==2.3.1

# Time Series & Financial Data
statsmodels==0.14.0
arch==6.2.0
fredapi==0.5.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Data Validation & Serialization
marshmallow==3.20.1
cerberus==1.3.5

# File Format Support
xlsxwriter==3.1.9
openpyxl==3.1.2
python-docx==1.1.0

# Networking & WebSocket
websocket-client==1.6.4
socket.io==0.2.1

# Async Support
asyncio-mqtt==0.13.0
aiofiles==23.2.1
asyncpg==0.29.0

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code Quality (Development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation (Development)
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Jupyter Notebooks (Development)
jupyter==1.0.0
jupyterlab==4.0.9
ipykernel==6.26.0

# Performance & Profiling
line-profiler==4.1.1
memory-profiler==0.61.0
py-spy==0.3.14

# Image Processing (for charts/graphs)
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Email & Notifications
sendgrid==6.10.0
twilio==8.10.3
slack-sdk==3.26.1

# Backup & Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Encryption & Hashing
bcrypt==4.1.1
argon2-cffi==23.1.0

# API Documentation
fastapi-users==12.1.2
fastapi-limiter==0.1.5

# Market Data APIs (Additional)
polygon-api-client==1.13.2
iex-api-python==0.0.6
quandl==3.7.0

# Time Series Forecasting
prophet==1.1.5
pmdarima==2.0.4

# Natural Language Processing
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2

# Caching
python-memcached==1.62
diskcache==5.6.3

# Configuration Management
dynaconf==3.2.4
omegaconf==2.3.0

# Process Management
supervisor==4.2.5
gunicorn==21.2.0

# System Monitoring
psutil==5.9.6
py-cpuinfo==9.0.0

# Development & Debugging Tools (Optional)
ipdb==0.13.13
pudb==2023.1
icecream==2.1.3

# Version Pinning for Stability
setuptools==69.0.2
wheel==0.42.0
pip==23.3.1

# Platform Specific (Windows - TA-Lib binary wheel)
# Note: Install TA-Lib manually using the provided wheel file
# ta_lib-0.6.3-cp313-cp313-win_amd64.whl